requirements.txt
Flask==3.1.0
keras==3.6.0
matplotlib==3.10.0
numpy==2.2.1
Pillow==11.0.0
scikit_learn==1.6.0
seaborn==0.13.2
tensorflow==2.18.0
tensorflow_intel==2.18.0